test(tool): satisfy complete_step repository budgets
Problem:
The PR's terminal-message coverage pushed completestep_test.go past the 800-line test ceiling, while two contract comments added five essay units beyond the existing per-file baseline.

Root cause:
The terminal-path tests were appended to an already near-limit test file, and the deliberate comments documenting final-todo orchestration were not reflected in the essay ratchet.

Fix:
Move the three terminal-path tests into completestep_terminal_test.go and raise only completestep.go's essay baseline from 3 to 8. Preserve the comments because they explain the termination contract and why the model must not continue.

Verification:
- go run ./tools/repolint
- go test ./internal/tool/builtin
- go test ./internal/agent -run TestE2ESerialPlanHostAdvancesAndAllowsFinalAnswer -count=1
- go vet ./internal/tool/builtin ./internal/agent
- git diff --check
