+++ tools/repolint/baseline.json
@@ -1883,7 +1883,7 @@
       "essay": 14
     },
     "internal/tool/builtin/completestep.go": {
-      "essay": 3
+      "essay": 8
     },
     "internal/tool/builtin/confine.go": {
       "essay": 16
